+++ ocrd/cli/workspace.py
@@ -114,9 +114,10 @@ def workspace_create(ctx, clobber_mets, directory):
 @click.option('-i', '--file-id', help="ID for the file", required=True)
 @click.option('-m', '--mimetype', help="Media type of the file", required=True)
 @click.option('-g', '--group-id', help="GROUPID")
+@click.option('--force', help="If file with ID already exists, replace it", default=False, is_flag=True)
 @click.argument('local_filename', type=click.Path(dir_okay=False, readable=True, resolve_path=True), required=True)
 @pass_workspace
-def workspace_add_file(ctx, file_grp, file_id, mimetype, group_id, local_filename):
+def workspace_add_file(ctx, file_grp, file_id, mimetype, group_id, force, local_filename):
     """
     Add a file to METS in a workspace.
     """
@@ -134,6 +135,7 @@ def workspace_add_file(ctx, file_grp, file_id, mimetype, group_id, local_filenam
         mimetype=mimetype,
         url=url,
         groupId=group_id,
+        force=force,
         local_filename=local_filename
     )
     workspace.save_mets()